Compress from www dir to fix build.
Before the full local path was contained in the ZIP-file and uploaded as such to PhoneGap Build. Due to a recent change in their machinery, their build script will not traverse the whole directory tree anymore but expects to find index.html in a top-level www directory or in the root of the ZIP-file.

Reference: http://community.phonegap.com/nitobi/topics/error_no_index_html_found-xcvob